refactor: Move all theming into a new folder in core
