+++ modules/common/react/lib/types.ts
@@ -1,5 +1,3 @@
-import {CanvasTheme} from '@workday/canvas-kit-react-core';
-
 /**
  * The growth behavior of a component.
  */
@@ -8,13 +6,6 @@ export interface GrowthBehavior {
    * True if the component should grow to its container's width. False otherwise.
    */
   grow?: boolean;
-}
-
-/**
- * Indicates a component is themeable with a CanvasTheme
- */
-export interface Themeable {
-  theme?: CanvasTheme;
 }
 
 /**
